Moves content between different files and adds my own icon
